redid a lot of bsp functions, flats, some other things as int16. it seems there is a render bug from an earlier commit visible in demo 3 where there are some weird vertical lines around when the 1st spectre is killed during the demo. probably related to a bad 32 to 16 bit conversion of a variable

+++ d_main.c
@@ -124,14 +124,14 @@ boolean         singletics = false; // debug flag to cancel adaptiveness
 
 
 
-extern  int32_t     sfxVolume;
-extern  int32_t     musicVolume;
+extern  uint8_t     sfxVolume;
+extern  uint8_t     musicVolume;
 
 extern  boolean inhelpscreens;
 
 skill_t         startskill;
-int32_t             startepisode;
-int32_t             startmap;
+int8_t             startepisode;
+int8_t             startmap;
 boolean         autostart;
 
 FILE*           debugfile;
