+++ p_inter.h
@@ -20,7 +20,7 @@
 #ifndef __P_INTER__
 #define __P_INTER__
 
-boolean	P_GivePower(player_t*, int32_t);
+boolean	P_GivePower(player_t*, int16_t);
